refactored color helpers
